Added tool to include updated description in Zenodo file
There is a description on the Zenodo file that must be provided in
HTML format. However, we want that description to replicate our
README file, that is written in Markdown. I have created a script that
reads the README file and introduces it into the description section
of the Zenodo json file. It generates a zenodo.json file in the utils
folder that needs to be manualy copied back to the base folder.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   utils/Zenodo_File_Template
	new file:   utils/update_zenodo_description.sh
	new file:   utils/zenodo.json